requirements.txt
streamlit==1.28.0
plotly==5.17.0
pandas==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.0
python-dotenv==1.0.0